% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_distance.R
\name{dist_forest_irf}
\alias{dist_forest_irf}
\title{Signed iterative random forest distance (dist_sirf)}
\usage{
dist_forest_irf(fit, x, n.core = 1)
}
\arguments{
\item{fit}{a fitted random forest of class ranger (probability = FALSE) or
randomForest.}

\item{x}{feature matrix, indicating observations to compute pairwise
distances for.}

\item{n.core}{number of workers to distribute readForest over.}
}
\value{
a numeric matrix indicating sirf-based distance over each pair of
observations in x.
}
\description{
Compute sirf-based distance between all pairs of observations passed through
a fitted random forest.
}
