% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forestDist}
\alias{forestDist}
\title{Computes distances between pairs of observations using iRF-encoded binary
features.}
\usage{
forestDist(obsn, featn, idtree, ridx, cidx)
}
\arguments{
\item{obsn}{node membership matrix indicating which leaf node observations
fall in for each tree, rows being observations, cols being trees.}

\item{featn}{binary matrix indicating iRF decision path encodings, rows
being nodes, cols being features.}

\item{idtree}{vector indicating rows of featn that correspond to new trees.}

\item{ridx}{first index of a pair for which to compute distance.}

\item{cidx}{second index of a pair for which to compute distance.}
}
\value{
Product of v1 and v2
}
\description{
Computes distances between pairs of observations using iRF-encoded binary
features.
}
